requirements.txt
baycomp==1.0.3
build==1.0.3
cmake==3.27.7
dataclasses==0.6
dp-accounting==0.4.3
eventfd==0.2
face-recognition==1.3.0
future==0.18.3
keras-nightly==2.5.0.dev2021032900
keras-preprocessing==1.1.2
lapsolver==1.1.0
libtmux==0.23.2
munch==4.0.0
opencv-python==4.8.1.78
pandas==2.1.2
patool==1.15.0
pip-search==0.0.12
pyarrow==14.0.1
pycryptodome==3.19.0
pyemd==1.0.0
tensorboard==2.15.1
torchac==0.9.3
torcheval==0.0.7
torchtext==0.16.0
torchvision==0.16.0
